requirements.txt
google-generativeai
langchain
chromadb
pytube
youtube-transcript-api
openai
pytorch
torchvision
ftfy
regex
tqdm
yt-dlp
langchain-chroma
langchain-community
langchain-huggingface
langchain-google-genai
ffmpeg-python
streamlit